Farm-ease
Indian Agriculture sector accounts for 18 per cent of   India’s  gross domestic product(GDP) and provides employment to 50% of the countries workforce. India is the world’s largest producer of pulse, rice, wheat, spices and spice product. Agriculture means the cultivation of crops for food and fodder. We feel gratitude when we hear the word “agriculture”. Without agriculture, it is not possible to feed ourselves. Our farmers stand for us in any situation by giving food to the world. But there is an harsh reality that farmer face problem in the  cultivation of crops. The seeds and fertilizer farmers need to grow their crops are just too expensive which they cannot afford for which they have to takes loans from the banks. During cultivation some of the crops are destroyed because of the diseases, water scarcity or unpredictable weather events.   Even a good harvest can be a double-edged sword, as farmers might not receive fair prices for their produce, further hindering their ability to repay loans. This combination of factors can lead to a crippling debt burden, pushing some farmers to the brink of despair. The issue goes beyond food security and touches upon the well-being of millions of individuals. So, farmers left with no other option rather than suicide.
To overcome this problem we introduce an idea, named “Farm-ease”, which will provide an interface to the farmer and help them to generate maximum profit for their crops. “Farm-ease” help to established an smart contract between farmer and consumer, this smart contract will be immutable which means once the contract is established it cannot tempered. And the smart contact will be finalized according to the terms and conditions mutually decided by the farmer and consumer.